Optimised space sections code
