+++ app/lib/features/space/pages/space_details_page.dart
@@ -1,12 +1,12 @@
 import 'package:acter/common/providers/room_providers.dart';
 import 'package:acter/features/space/providers/space_navbar_provider.dart';
-import 'package:acter/features/space/widgets/space_details/about_section.dart';
-import 'package:acter/features/space/widgets/space_details/chats_section.dart';
-import 'package:acter/features/space/widgets/space_details/events_section.dart';
-import 'package:acter/features/space/widgets/space_details/members_section.dart';
-import 'package:acter/features/space/widgets/space_details/pins_section.dart';
-import 'package:acter/features/space/widgets/space_details/spaces_sesction.dart';
-import 'package:acter/features/space/widgets/space_details/tasks_section.dart';
+import 'package:acter/features/space/widgets/space_sections/about_section.dart';
+import 'package:acter/features/space/widgets/space_sections/chats_section.dart';
+import 'package:acter/features/space/widgets/space_sections/events_section.dart';
+import 'package:acter/features/space/widgets/space_sections/members_section.dart';
+import 'package:acter/features/space/widgets/space_sections/pins_section.dart';
+import 'package:acter/features/space/widgets/space_sections/spaces_sesction.dart';
+import 'package:acter/features/space/widgets/space_sections/tasks_section.dart';
 import 'package:acter/features/space/widgets/space_header.dart';
 import 'package:acter/features/space/widgets/space_toolbar.dart';
 import 'package:animated_size_and_fade/animated_size_and_fade.dart';
