--- app/lib/features/space/widgets/space_details/pins_section.dart
@@ -1,74 +1,0 @@
-import 'package:acter/common/providers/space_providers.dart';
-import 'package:acter/common/toolkit/buttons/inline_text_button.dart';
-import 'package:acter/features/pins/providers/pins_provider.dart';
-import 'package:acter/features/pins/widgets/pin_list_item.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_gen/gen_l10n/l10n.dart';
-
-class PinsSection extends ConsumerWidget {
-  final String spaceId;
-  final int limit;
-
-  const PinsSection({
-    super.key,
-    required this.spaceId,
-    this.limit = 3,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        pinsLabel(context),
-        pinsList(context, ref),
-      ],
-    );
-  }
-
-  Widget pinsLabel(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 8),
-      child: Row(
-        children: [
-          Text(
-            L10n.of(context).pins,
-            style: Theme.of(context).textTheme.titleMedium,
-          ),
-          const Spacer(),
-          ActerInlineTextButton(
-            onPressed: () {},
-            child: Text(L10n.of(context).seeAll),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget pinsList(BuildContext context, WidgetRef ref) {
-    final space = ref.watch(spaceProvider(spaceId)).requireValue;
-    final pins = ref.watch(spacePinsProvider(space));
-
-    return pins.when(
-      data: (pins) {
-        int pinsLimit = (pins.length > limit) ? limit : pins.length;
-        return ListView.builder(
-          shrinkWrap: true,
-          itemCount: pinsLimit,
-          padding: EdgeInsets.zero,
-          physics: const NeverScrollableScrollPhysics(),
-          itemBuilder: (context, index) {
-            return PinListItemById(pinId: pins[index].eventIdStr());
-          },
-        );
-      },
-      error: (error, stack) =>
-          Center(child: Text(L10n.of(context).loadingFailed(error))),
-      loading: () => Center(
-        child: Text(L10n.of(context).loading),
-      ),
-    );
-  }
-}
