+++ app/lib/features/space/widgets/space_sections/tasks_section.dart
@@ -1,0 +1,63 @@
+import 'package:acter/features/space/widgets/space_sections/section_header.dart';
+import 'package:acter/features/tasks/providers/tasklists.dart';
+import 'package:acter/features/tasks/widgets/task_list_item_card.dart';
+import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_gen/gen_l10n/l10n.dart';
+
+class TasksSection extends ConsumerWidget {
+  final String spaceId;
+  final int limit;
+
+  const TasksSection({
+    super.key,
+    required this.spaceId,
+    this.limit = 3,
+  });
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final taskList = ref.watch(spaceTasksListsProvider(spaceId));
+    return taskList.when(
+      data: (tasks) => buildTasksSectionUI(context, tasks),
+      error: (error, stack) =>
+          Center(child: Text(L10n.of(context).loadingFailed(error))),
+      loading: () => Center(
+        child: Text(L10n.of(context).loading),
+      ),
+    );
+  }
+
+  Widget buildTasksSectionUI(BuildContext context, List<TaskList> tasks) {
+    int taskLimit = (tasks.length > limit) ? limit : tasks.length;
+    bool isShowSeeAllButton = tasks.length > taskLimit;
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.start,
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        SectionHeader(
+          title: L10n.of(context).tasks,
+          isShowSeeAllButton: isShowSeeAllButton,
+          onTapSeeAll: () {},
+        ),
+        taskListUI(tasks, taskLimit),
+      ],
+    );
+  }
+
+  Widget taskListUI(List<TaskList> tasks, int taskLimit) {
+    return ListView.builder(
+      shrinkWrap: true,
+      itemCount: taskLimit,
+      padding: EdgeInsets.zero,
+      physics: const NeverScrollableScrollPhysics(),
+      itemBuilder: (context, index) {
+        return TaskListItemCard(
+          taskList: tasks[index],
+          initiallyExpanded: false,
+        );
+      },
+    );
+  }
+}
